Add files via upload
Relational Databases in Python
Data Wrangling and Relational Databases
In the context of data wrangling, we recommend that databases and SQL only come into play for gathering data or storing data. That is:

Connecting to a database and importing data into a pandas DataFrame (or the analogous data structure in your preferred programming language), then assessing and cleaning that data, or
Connecting to a database and storing data you just gathered (which could potentially be from a database), assessed, and cleaned
These tasks are especially necessary when you have large amounts of data, which is where SQL and other databases excel over flat files.

The two scenarios above can be further broken down into three main tasks:

Connecting to a database in Python
Storing data from a pandas DataFrame in a database to which you're connected, and
Importing data from a database to which you're connected to a pandas DataFrame
This Lesson
For the example in this lesson, we're going to do these in order:

Connect to a database. We'll connect to a SQLite database using SQLAlchemy, a database toolkit for Python.
Store the data in the cleaned master dataset in that database. We'll do this using pandas' to_csv DataFrame method.
Then read the brand new data in that database back into a pandas DataFrame. We'll do this using pandas' read_csv function.
The third one isn’t necessary for this lesson, but often in the workplace, instead of having to download files, scrape web pages, hit an API, etc., you're given a database right at the beginning of a project.

All three of these tasks will be introduced and carried out in the Jupyter Notebook below. These are not quizzes. All of the code is provided for you. Your job is to read and understand each comment and line of code, then run the code.